Отрисовывает комментарии по данным сервера
