+++ js/api.js
@@ -8,7 +8,7 @@ const getData = () => fetch('https://28.javascript.pages.academy/kekstagram/data
   });
 
 const sendData = (body) => fetch(
-  'https://28.javascript.pages.academy/kekstagra',
+  'https://28.javascript.pages.academy/kekstagram',
   {
     method: 'POST',
     body,
